Fix a bug of repeated downloading images

+++ qiushibaike.py
@@ -94,7 +94,9 @@ def insert(date, content, image="", location=""):
 def download(url, path):
     filename = re.split('/', urlparse.urlparse(url).path)[-1]
     filepath = os.path.join(path, filename)
-    urllib.urlretrieve(url, filepath)
+    
+    if os.path.isfile(filepath):
+        urllib.urlretrieve(url, filepath)
 
     return filename
